Clone Upgrade Version 1.2.0
Here is the "Prestige " of the Magic "Job Meta Search Engine".
This Script take Job Title and Location as an Input.
Goes toward all 4 APIs, Retrive Results from their.
Sort all 4 APIs Result According to time of posting. And display it.
Here The Indeed API key has been Removed. Register on the indeed.com and have your Own key.

This script take 20 days to be completed. So respect it. These one is the spoon-feed for you guys. Cause I don't want you to face the problems that I did.
So Enjoy the Script.